refactor: Convert Audit Module to Typescript
