+++ src/app/core/audit/audit.controller.ts
@@ -1,20 +1,16 @@
-'use strict';
-
-const _ = require('lodash'),
-	deps = require('../../../dependencies'),
-	dbs = deps.dbs,
-	util = deps.utilService,
-	Audit = dbs.admin.model('Audit');
+import _ from 'lodash';
+import { dbs, utilService as util } from '../../../dependencies';
+const Audit = dbs.admin.model('Audit');
 
 /**
  * Retrieves the distinct values for a field in the Audit collection
  */
-module.exports.getDistinctValues = async (req, res) => {
+export const getDistinctValues = async (req, res) => {
 	const results = await Audit.distinct(req.query.field, {});
 	res.status(200).json(results);
 };
 
-module.exports.search = async function (req, res) {
+export const search = async function (req, res) {
 	const search = req.body.s || null;
 	let query = req.body.q || {};
 	query = util.toMongoose(query);
